backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Plot Platform API - コミュニティ企画プラットフォーム"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "supabase",
    "python-multipart",
    "Pillow",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-dotenv",
    "asyncpg",
    "psycopg2-binary",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
